refactor: move focus and close methods to pane grid
